+++ ki/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class KiConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "ki"
+    verbose_name = "Modell-Steckplatz"