+++ parameter/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class ParameterConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "parameter"
+    verbose_name = "Parameterregister"